calends/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
gunicorn==21.2.0
idna==3.6
lxml==4.9.3
packaging==23.2
python-docx==1.1.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
